+++ threatlens/__init__.py
@@ -5,5 +5,5 @@ from importlib.metadata import PackageNotFoundError, version
 try:
     __version__ = version("threatlens-cli")
 except PackageNotFoundError:
-    __version__ = "2.2.1"
+    __version__ = "2.3.0"
 __author__ = "Jude Hilgendorf"